Fix bad double to integer cast in mixer with bad loop values.

+++ src/mixer.c
@@ -476,7 +476,7 @@ void libxmp_mixer_softmixer(struct context_data *ctx)
 				samples = 0;
 				usmp = 1;
 			} else {
-				int c = ceil(((double)vi->end - vi->pos) / step);
+				double c = ceil(((double)vi->end - vi->pos) / step);
 				/* ...inside the tick boundaries */
 				if (c > size) {
 					c = size;
